data_ingestion/individual-downloaders/requirements.txt
earthengine-api
pandas
pyarrow
google-cloud-storage
